No more false int64 support

+++ vm/Operand.cpp
@@ -5,9 +5,9 @@ namespace exec {
 template <>
 IOperand::OperandType Operand<int8_t>::s_type = IOperand::OperandType::Int8;
 template <>
-IOperand::OperandType Operand<int32_t>::s_type = IOperand::OperandType::Int16;
+IOperand::OperandType Operand<int16_t>::s_type = IOperand::OperandType::Int16;
 template <>
-IOperand::OperandType Operand<int64_t>::s_type = IOperand::OperandType::Int32;
+IOperand::OperandType Operand<int32_t>::s_type = IOperand::OperandType::Int32;
 template <>
 IOperand::OperandType Operand<float>::s_type = IOperand::OperandType::Float;
 template <>
